+++ HISTORY.md
@@ -5,6 +5,7 @@
 - feat: small accouncement feature for players going to bed (default off) [0.11.2]
 - refactor: moved onClick from `BaseButton` to `OZUIElement` [0.11.1]
 - refactor: streamlined welcome message property [0.11.2]
+- refactor: fix invisible line chat spam [0.11.3]
 
 ## [0.11.0] - 2025-12-18 | OZ UI