+++ HISTORY.md
@@ -6,6 +6,7 @@
 - refactor: moved onClick from `BaseButton` to `OZUIElement` [0.11.1]
 - refactor: streamlined welcome message property [0.11.2]
 - refactor: fix invisible line chat spam [0.11.3]
+- refactor: wrong player language for bed message [0.11.4]
 
 ## [0.11.0] - 2025-12-18 | OZ UI