+++ HISTORY.md
@@ -2,6 +2,9 @@
 
 ## [unreleased]
 
+## [0.12.0] - 2025-12-24 | PlayerSettings
+
+- feat: `PlayerSettings` database to persist player-plugin-settings [0.12.0]
 - feat: small accouncement feature for players going to bed (default off) [0.11.2]
 - refactor: moved onClick from `BaseButton` to `OZUIElement` [0.11.1]
 - refactor: streamlined welcome message property [0.11.2]